+++ lib/repositories/artist.repository.ts
