+++ lib/repositories/index.ts
@@ -1,1 +1,5 @@
 export * from './base.repository';
+export * from './album.repository';
+export * from './artist.repository';
+export * from './track.repository';
+export * from './user.repository';
