+++ src/album/album.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TrackModule } from 'src/track/track.module';
 import { AlbumController } from './album.controller';
-import { AlbumRepository } from './album.repository';
+import { AlbumRepository } from '../../lib/repositories';
 import { AlbumService } from './album.service';
 
 @Module({
