+++ src/artist/artist.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { AlbumModule } from 'src/album/album.module';
 import { TrackModule } from 'src/track/track.module';
 import { ArtistController } from './artist.controller';
-import { ArtistRepository } from './artist.repository';
+import { ArtistRepository } from '../../lib/repositories';
 import { ArtistService } from './artist.service';
 
 @Module({
