+++ src/user/user.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { UserController } from './user.controller';
-import { UserRepository } from './user.repository';
+import { UserRepository } from '../../lib/repositories';
 import { UserService } from './user.service';
 
 @Module({
